splitting page into components
