+++ client/src/App.js
@@ -1,16 +1,21 @@
-import React, { Component } from "react";
-import { BrowserRouter, Route, Switch, Redirect } from "react-router-dom";
-import "./App.css";
-import { library } from "@fortawesome/fontawesome-svg-core";
-import { faTrash, faFilter, faPlus, faCaretDown } from "@fortawesome/free-solid-svg-icons";
-import Login from "./components/Login";
-import Header from "./components/Layout/Header";
-import Projects from "./components/Projects";
-import ProjectNew from "./components/Projects/ProjectNew";
-import ProjectEdit from "./components/Projects/ProjectEdit";
-import Members from "./components/Members";
-import Scrum from "./components/scrum";
-import { SearchIcon } from "./components/common components/search";
+import React, { Component } from 'react';
+import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
+import './App.css';
+import { library } from '@fortawesome/fontawesome-svg-core';
+import {
+  faTrash,
+  faFilter,
+  faPlus,
+  faCaretDown
+} from '@fortawesome/free-solid-svg-icons';
+import Login from './components/Login';
+import Header from './components/Layout/Header';
+import Projects from './components/Projects';
+import ProjectNew from './components/Projects/ProjectNew';
+import ProjectEdit from './components/Projects/ProjectEdit';
+import Members from './components/Members';
+import Scrum from './components/scrum';
+import { SearchIcon } from './components/commonComponents/search';
 
 library.add(faTrash, faFilter, faPlus, faCaretDown);
