+++ client/src/components/commonComponents/search/style.css
